Move media index changing to proc_handler
If we allow changing of the media index using settings updating,
clicking Cancel on the Properties window causes the source to play the
file that was being played at the time that the properties window was
opened, which is very confusing, especially if the user doesn't even
need to set the media index programmatically. Would still be a problem
if we decide to enabled deferred updates of settings.