Add Speed option
Media Source restarts when speed is modified so we have to add such
warning.
We have to disable refreshing properties on settings update because it
will refresh while dragging the speed slider, making it almost
impossible to use it. Thus we also add a button to refresh the current
filename.